Update - In AnalysisMethods: Bugs bei den beiden prod()-Methoden behoben.

+++ src/abstractexpressions/expression/classes/Operator.java
@@ -1714,7 +1714,7 @@ public class Operator extends Expression {
                     BigInteger.valueOf(ComputationBounds.BOUND_OPERATOR_MAX_NUMBER_OF_MEMBERS_IN_SUM_OR_PRODUCT)) <= 0) {
 
                 // Dann kann man die Summe explizit ausschreiben.
-                return AnalysisMethods.prod(factor, (String) this.params[1], lowerLimit.intValue(), upperLimit.intValue());
+                return AnalysisMethods.prod(factor, (String) this.params[1], lowerLimit, upperLimit);
 
             }
         }
